move utils, send email if error
